requirements.txt
asgiref==3.9.1
channels==4.2.2
channels_redis==4.3.0
Django==5.2.4
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
msgpack==1.1.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
redis==6.2.0
sqlparse==0.5.3
tzdata==2025.2
